refactor: move tests for cogs
