+++ tests/test_cogs/test_Announce.py
