+++ tests/test_cogs/test_ColourRole.py
