+++ tests/test_cogs/test_ReactForRole.py
