+++ tests/test_cogs/test_TextFilter.py
