+++ tests/test_cogs/test_Voting.py
